implemented the logic in the table view controller
